rename module recognizers into input_processor
