+++ presenter/src/input_processor/pan.rs
@@ -1,4 +1,4 @@
-use crate::recognizer::transaction::Transaction;
+use crate::input_processor::transaction::Transaction;
 use crate::{InputProcessor, InputState};
 use emergent_drawing::Point;
 use emergent_ui::{WindowEvent, WindowMessage};
