+++ presenter/src/input_processor/subscriptions.rs
